Simulate a Micro array data set with accompanying metadata

Simulate Micro array data by randomly drawing values from a normal distribution centered at 10 with a standard deviation of 5

MakeMArrayDataSet(n_samples = 10L, n_features = 6L, with_seed = 8675309)

Arguments

n_samples

The number of samples to simulate (must be a positive, even integer value that is no greater than '26')

n_features

The number of genes to simulate (must be a positive, even integer value)

with_seed

A numeric input to set the seed of the session (defaults to '8675309')

Value

A list of 1 matrix ('marray' - simulated metadata), and 2 data frames (1 for sample metadata -- 'pmeta' -- and another for gene metadata -- 'fmeta')

Examples

MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
#> $marray
#>                 A         B         C         D         E         F         G
#> probe1  14.248861  4.882717 13.021513  4.883014 19.843221 12.897883  6.760087
#> probe2   4.819262 14.168907  9.501717 22.081590 14.878788 11.977335 10.110728
#> probe3   6.215572 16.262428 13.097487  9.660019  2.063335 11.804046  7.578993
#> probe4  14.510729 16.063694 15.395166 13.425103  8.297149  9.746559 10.931769
#> probe5  17.917186  2.209321  9.985245 14.048307  9.750469  8.369924  9.116786
#> probe6  14.173852 12.725705 11.952293  8.686001  7.328275 11.569101  3.899764
#> probe7  13.484770 17.796806 14.538239  4.451559  8.486790  9.447696 12.428832
#> probe8   8.568416 11.897215  5.296738 -3.122392  5.850877 11.398446 10.070544
#> probe9   8.397948 15.672457 14.541579 13.155194  3.200742 10.702447  6.297323
#> probe10  7.863364  6.400357  3.360809 11.894879  6.484152  5.482783  4.518485
#> probe11  8.105213 16.928951  5.428222  7.585015  9.288727 18.839257  3.961754
#> probe12 13.991144 10.115327 16.728921  8.599433  4.287773  2.602798 14.130592
#>                  H         I          J         K          L         M
#> probe1   7.3076041 13.482314 -0.9229271 12.835828  7.1797326  8.821433
#> probe2   8.4577991 14.169480 12.6863777 17.276546  6.3591267 -3.201893
#> probe3  14.7488169 15.911143  7.7498663 14.190946  1.0455111 -1.550707
#> probe4   8.5619253 15.961283 20.6459410  7.916914  7.8680591 11.128538
#> probe5  13.7288917  7.566246  7.3780254  9.763899  9.9249592 10.448715
#> probe6   7.4498186 10.993117  1.3105838 10.700182  5.7128758 13.718445
#> probe7  -0.6282565  8.260899  9.0477887 12.111686 17.8504647  6.866545
#> probe8   6.1349583 10.679594  9.5803330  9.433018 10.8981485 16.797578
#> probe9  13.0722424 10.800073 10.4777834  6.916787 21.9485822 11.579145
#> probe10 10.6124438  9.261950 10.0893342  4.058576 -0.8801124  9.130950
#> probe11  9.8865256  3.872160  6.4021636 13.217448  8.8280074 17.226016
#> probe12 11.4467985  6.583292 14.9558544 19.361845  6.0322631 11.907153
#>                 N         O         P
#> probe1  18.484418 13.300045 10.634071
#> probe2  10.727246 -2.534951 -1.396437
#> probe3  16.164681  5.139030 11.025951
#> probe4   1.203143 17.507647 11.726380
#> probe5   7.550686  9.158758 11.465710
#> probe6  11.671338 16.844544  6.906901
#> probe7  20.247732 11.284889  6.540971
#> probe8   4.384174  9.121898  3.860821
#> probe9  15.685482  6.652354 11.242889
#> probe10  8.565395 12.559918 -1.497512
#> probe11  4.574426 12.881198 12.461880
#> probe12 13.621260  5.567145  9.940492
#> 
#> $fmeta
#>         geneId pathway
#> probe1       1       M
#> probe2       2       C
#> probe3       3       Y
#> probe4       4       O
#> probe5       5       G
#> probe6       6       X
#> probe7       7       Z
#> probe8       8       Z
#> probe9       9       T
#> probe10     10       F
#> probe11     11       C
#> probe12     12       P
#> 
#> $pmeta
#>   sampleId condition
#> A        1        WT
#> B        2        WT
#> C        3        WT
#> D        4        WT
#> E        5        WT
#> F        6        WT
#> G        7        WT
#> H        8        WT
#> I        9       MUT
#> J       10       MUT
#> K       11       MUT
#> L       12       MUT
#> M       13       MUT
#> N       14       MUT
#> O       15       MUT
#> P       16       MUT
#>