boxplot {marray}	R Documentation

Boxplots for cDNA microarray spot statistics

Description

The function boxplot produces boxplots of microarray spot statistics for the classes "marrayRaw", "marrayNorm". We encourage users to use boxplot rather than maBoxplot. The name of the arguments have changed slightly.

Usage


## S4 method for signature 'marrayRaw':
boxplot(x, xvar="maPrintTip", yvar="maM", ...)
## S4 method for signature 'marrayNorm':
boxplot(x, xvar="maPrintTip", yvar="maM", ...)

Arguments

x 	 Microarray object of class "marrayRaw", "marrayNorm"

xvar 	 Name of accessor method for the spot statistic used to stratify the data, typically a slot name for the microarray layout object (see "marrayLayout") such as maPlate or a method such as maPrintTip. If x is NULL, the data are not stratified.

yvar 	 Name of accessor method for the spot statistic of interest, typically a slot name for the microarray object m, such as maM.

... 	 Optional graphical parameters, see par.

Details

If there are more than one array in the batch, the function produces a boxplot for each array in the batch. Such plots are useful when assessing the need for between array normalization, for example, to deal with scale differences among different arrays. Default graphical parameters are chosen for convenience using the function maDefaultPar (e.g. color palette, axis labels, plot title) but the user has the option to overwrite these parameters at any point.

Author(s)

Jean Yang and Sandrine Dudoit

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maBoxplot, maDefaultPar.

Examples



# To see the demo type demo(marrayPlots)

# Examples use swirl dataset, for description type ? swirl
data(swirl)

# Boxplots of pre-normalization log-ratios M for each of the 16 
# print-tip-groups for the Swirl 93 array. 
# - Default arguments
boxplot(swirl[,3])  

# All spots 
boxplot(swirl[,3], xvar=NULL, col="green")  

# Boxplots of pre-normalization red foreground intensities for each grid row
# for the Swirl 81 array. 
boxplot(swirl[,1], xvar="maGridRow", yvar = "maRf", main = "Swirl array 81: pre-normalization red foreground intensity") 

# Boxplots of pre-normalization log-ratios for each array in swirl
boxplot(swirl, main="Swirl arrays: pre-normalization log-ratios")

[Package marray version 1.18.0 Index]