[-methods {marray}	R Documentation

Subsetting methods for microarray objects

Description

Subsetting methods were defined for the microarray classes, marrayInfo, marrayLayout,marrayRaw and marrayNorm. These methods create instances of the given class, for a subset of spots and/or arrays in a batch.

Methods

x = ANY

generic method.

x = marrayInfo

x[i, j] extract object of class "marrayInfo" for spots or arrays with indices i and labels with indices j.

x = marrayLayout

x[i] extract object of class "marrayLayout" for spots with indices i.

x = marrayRaw

x[i, j] extract object of class "marrayRaw" for spots with indices i and arrays with indices j.

x = marrayNorm

x[i, j] extract object of class "marrayNorm" for spots with indices i and arrays with indices j.

[Package marray version 1.18.0 Index]