cbind {marray}	R Documentation

Combine marrayRaw, marrayNorm or marrayInfo Objects

Description

Combine a series of marrayRaw, marrayNorm and marrayInfo objects.

Usage


## S3 method for class 'marrayRaw':
cbind(..., deparse.level=1)
## S3 method for class 'marrayNorm':
cbind(..., deparse.level=1)
## S3 method for class 'marrayInfo':
rbind(..., deparse.level=1)

Arguments

... 	 marrayRaw objects or marrayNorm objects

deparse.level 	 not currently used, see cbind in the base package

Details

cbind combines data objects assuming the same gene lists but different arrays. rbind combines data objects assuming equivalent arrays, i.e., the same RNA targets, but different genes.

For cbind, the matrices o f expression data from the individual objects are cbinded. The data.frames of target information, if they exist, are rbinded. The combined data object will preserve any additional components or attributes found in the first object to be combined. For rbind, the matrices of expression data are rbinded while the target information, in any, is unchanged.

Author(s)

Jean Yang

See Also

cbind in the base package.

[Package marray version 1.18.0 Index]