checkTargetInfo {marray}	R Documentation

Verifying the order between intensities matrix and target file information

Description

Check that the foreground and backgruond intensities are stored in the same order as provided in the first column of target file.

Usage


checkTargetInfo(mraw)

Arguments

mraw 	 Object of class marrayRaw or marryNorm.

Value

A logical value. This function returns "TRUE" if the first column from the Target information is the same order as the foreground and backgruond intensities.

Author(s)

Yee Hwa (Jean) Yang

Examples


datadir <- system.file("swirldata", package="marray")
swirl.targets <- read.marrayInfo(file.path(datadir, "SwirlSample.txt"))
data(swirl)
swirl@maTargets <- swirl.targets

checkTargetInfo(swirl)

checkTargetInfo(swirl[, 2:4])

## reorder
swirl@maTargets <- swirl.targets[c(2:4, 1),]
checkTargetInfo(swirl)

[Package marray version 1.18.0 Index]