coerce-methods {marray}	R Documentation

Coerce an object to belong to a given microarray class

Description

Coercing methods were defined to convert microarray objects of one class into objects of another class, e.g., instances of the "marrayRaw" class into instances of the "marrayNorm" class.

Methods

from = marrayRaw, to = marrayNorm

convert an object of class "marrayRaw" into an object of class "marrayNorm".

Note

Use Package convert to convert object to other data types such as ExpressionSet and MAList.

[Package marray version 1.18.0 Index]