dim {marray}	R Documentation

Retrieve the Dimensions of an marrayRaw, marrayNorm or marrayInfo Object

Description

Retrieve the number of rows (genes) and columns (arrays) for an marrayRaw, marrayNorm or marrayInfo object.

Usage


## S3 method for class 'marrayRaw':
dim(x)

Arguments

x 	 an object of class marrayRaw, marrayNorm or marrayInfo

Details

Microarray data objects share many analogies with ordinary matrices in which the rows correspond to spots or genes and the columns to arrays. These methods allow one to extract the size of microarray data objects in the same way that one would do for ordinary matrices.

A consequence is that row and column commands nrow(x), ncol(x) and so on also work.

Value

Numeric vector of length 2. The first element is the number of rows (genes) and the second is the number of columns (arrays).

Author(s)

modified from Gordon Smyth's function

See Also

dim in the base package.

Examples


M <- A <- matrix(11:14,4,2)
rownames(M) <- rownames(A) <- c("a","b","c","d")
colnames(M) <- colnames(A) <- c("A1","A2")
MA <- new("marrayNorm", maM=M,maA=A)
dim(MA)
dim(M)

[Package marray version 1.18.0 Index]