findID {marray}	R Documentation

Find ID when given an accession number

Description

Search gene ID with a vector of accession number from gene names or ID values.

Usage


findID(text, Gnames = gnames, ID = "Name")

Arguments

text 	 A character strings of gene names or id names.

Gnames 	 An objects of marrayRaw, marrayNorm, ExpressionSet or data.frame of gene names information.

ID 	 The column of ID corresponding to 'text'.

Value

A numeric vector the gene ID.

Author(s)

Yee Hwa (Jean) Yang

See Also

grep

Examples


data(swirl)
findID("fb24a09", swirl, ID="ID")
findID("geno1", swirl)

[Package marray version 1.18.0 Index]