image {marray}	R Documentation

Color image for cDNA microarray spot statistics

Description

We encourage users calling "image" rather than "maImage". The name of the arguments are change slightly. The function image creates spatial images of shades of gray or colors that correspond to the values of a statistic for each spot on the array. The statistic can be the intensity log-ratio M, a spot quality measure (e.g. spot size or shape), or a test statistic. This function can be used to explore whether there are any spatial effects in the data, for example, print-tip or cover-slip effects.

Usage


## S4 method for signature 'marrayRaw':
image(x, xvar = "maM", subset = TRUE, col, contours=FALSE,  bar = TRUE, overlay=NULL, ol.col=1, colorinfo=FALSE, ...)
## S4 method for signature 'marrayNorm':
image(x, xvar = "maM", subset = TRUE, col, contours=FALSE,  bar = TRUE, overlay=NULL, ol.col=1, colorinfo=FALSE, ...)

Arguments

x 	 Microarray object of class "marrayRaw", "marrayNorm"

xvar 	 Name of accessor function for the spot statistic of interest, typically a slot name for the microarray object x, such as maM.

subset 	 A "logical" or "numeric" vector indicating the subset of spots to display on the image.

col 	 List of colors such as that generated by rainbow, heat.colors, topo.colors, terrain.colors, or similar functions. In addition to these color palette functions, a new function maPalette was defined to generate color palettes from user supplied low, middle, and high color values.

contours 	 If contours=TRUE, contours are plotted, otherwise they are not shown.

bar 	 If bar=TRUE, a calibration color bar is shown to the right of the image.

overlay 	 A logical vector of spots to be highlighted on the image plots.

ol.col 	 Color of the overlay spots.

colorinfo 	 A logical value indicating whether the function should return the color scale information.

... 	 Optional graphical parameters, see par.

Details

This function calls the general function maImage.func, which is not specific to microarray data. If there are more than one array in the batch, the plot is done for the first array, by default. Default color palettes were set for different types of spot statistics using the maPalette function. When x=c("maM", "maMloc", "maMscale"), a green-to-red color palette is used. When x=c("maGb", "maGf", "maLG"), a white-to-green color palette is used. When x=c("maRb", "maRf", "maLR"), a white-to-red color palette is used. The user has the option to overwrite these parameters at any point.

Value

If colorinfo is set to TRUE, the following list with elements will be returned.

x.col 	 vector of colors to be used for calibration color bar.

x.bar 	 vector of values to be used for calibration color bar.

summary 	 six number summary of the spot statistics, from the function summary.

Author(s)

Jean Yang and Sandrine Dudoit

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maImage, maImage.func, maColorBar, maPalette

Examples


# Examples use swirl dataset, for description type ? swirl
data(swirl)

# Microarray color palettes
Gcol <- maPalette(low = "white", high = "green", k = 50)
Rcol <- maPalette(low = "white", high = "red", k = 50)
BYcol <- maPalette(low = "blue", mid="gray", high = "yellow", k = 50)

# Color images of green and red background and foreground intensities
##image(swirl[, 2], xvar ="maGb")
##image(swirl[, 2], xvar ="maGf", subset = TRUE, col = Gcol, contours = FALSE, bar = TRUE, main="Swirl array 93")
##image(swirl[, 1], xvar ="maRb", contour=TRUE)
##image(swirl[, 4], xvar ="maRf", bar=FALSE)

# Color images of pre-normalization intensity log-ratios
##image(swirl[, 1])

# Color images with overlay spots
##image(swirl[, 3], xvar = "maA", overlay = maTop(maA(swirl[, 3]), h = 0.1, l = 0.1), bar = TRUE, main = "Image of A values with % 10 tails highlighted") 

# Color image of print-tip-group
##image(swirl[, 1],xvar = "maPrintTip")


[Package marray version 1.18.0 Index]