ma2D {marray}	R Documentation

Stratified bivariate robust local regression

Description

This function performs robust local regression of a variable z on predictor variables x and y, separately within values of a fourth variable g. It is used by maNorm2D for 2D spatial location normalization.

Usage


ma2D(x, y, z, g, w=NULL, subset=TRUE, span=0.4, ...)

Arguments

x 	 A numeric vector of predictor variables.

y 	 A numeric vector of predictor variables.

z 	 A numeric vector of responses.

g 	 Variables used to stratify the data.

w 	 An optional numeric vector of weights.

subset 	 A "logical" or "numeric" vector indicating the subset of points used to compute the fits.

span 	 The argument span which controls the degree of smoothing in the loess function.

... 	 Misc arguments

Details

z is regressed on x and y, separately within values of g using the loess function.

Value

A numeric vector of fitted values.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maNormMain, maNorm2D, loess.

Examples


# See examples for maNormMain.

[Package marray version 1.18.0 Index]