maBoxplot {marray}	R Documentation

Boxplots for cDNA microarray spot statistics

Description

The function maBoxplot produces boxplots of microarray spot statistics for the classes marrayRaw and marrayNorm.We encourage users to use "boxplot" rather than "maBoxplot". The name of the arguments have changed.

Usage


maBoxplot(m, x="maPrintTip", y="maM", ...)

Arguments

m 	 Microarray object of class "marrayRaw" and "marrayNorm"

x 	 Name of accessor method for the spot statistic used to stratify the data, typically a slot name for the microarray layout object (see "marrayLayout") such as maPlate or a method such as maPrintTip. If x is NULL, the data are not stratified.

y 	 Name of accessor method for the spot statistic of interest, typically a slot name for the microarray object m, such as maM.

... 	 Optional graphical parameters, see par.

Details

If there are more than one array in the batch, the function produces a boxplot for each array in the batch. Such plots are useful when assessing the need for between array normalization, for example, to deal with scale differences among different arrays. Default graphical parameters are chosen for convenience using the function maDefaultPar (e.g. color palette, axis labels, plot title) but the user has the option to overwrite these parameters at any point.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

boxplot, maDefaultPar.

Examples


## see example in boxplot

[Package marray version 1.18.0 Index]