maColorBar {marray}	R Documentation

Calibration bar for color images

Description

This function produces a color image (color bar) which can be used for the legend to another color image obtained from the functions image, maImage, or maImage.func.

Usage


maColorBar(x, horizontal=TRUE, col=heat.colors(50), scale=1:length(x), k=10, ...)

Arguments

x 	 If "numeric", a vector containing the "z" values in the color image, i.e., the values which are represented in the color image. Otherwise, a "character" vector representing colors.

horizontal 	 If TRUE, the values of x are represented as vertical color strips in the image, else, the values are represented as horizontal color strips.

col 	 Vector of colors such as that generated by rainbow, heat.colors, topo.colors, terrain.colors, or similar functions. In addition to these color palette functions, a new function maPalette was defined to generate color palettes from user supplied low, middle, and high color values.

scale 	 A "numeric" vector specifying the "z" values in the color image. This is used when the argument x is a "character" vector representing color information.

k 	 Object of class "numeric", for the number of labels displayed on the bar.

... 	 Optional graphical parameters, see par.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine, Yee Hwa (Jean) Yang.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

image, maImage, maImage.func, maPalette.

Examples


par(mfrow=c(3,1))
Rcol <- maPalette(low="white", high="red", k=10)
Gcol <- maPalette(low="white", high="green", k=50)
RGcol <- maPalette(low="green", high="red", k=100)
maColorBar(Rcol)
maColorBar(Gcol, scale=c(-5,5))
maColorBar(1:50, col=RGcol)

par(mfrow=c(1,3))
x<-seq(-1, 1, by=0.01)
maColorBar(x, col=Gcol, horizontal=FALSE, k=11)
maColorBar(x, col=Gcol, horizontal=FALSE, k=21)
maColorBar(x, col=Gcol, horizontal=FALSE, k=51)

[Package marray version 1.18.0 Index]