maCompCoord {marray}	R Documentation

Generate grid and spot matrix coordinates

Description

This function generates grid and spot matrix coordinates from ranges of rows and columns for the grid and spot matrices. Spots on the array are numbered consecutively starting from the top left grid and the top left spot within each grid.

Usage


maCompCoord(grows, gcols, srows, scols)

Arguments

grows 	 numeric vector of grid rows.

gcols 	 numeric vector of grid columns.

srows 	 numeric vector of spot rows.

scols 	 numeric vector of spot columns.

Value

a matrix of spot four-coordinates, with rows corresponding to spots and columns to grid row, grid column, spot row, and spot column coordinates.

Author(s)

Yee Hwa (Jean) Yang, Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

See Also

marrayLayout, maCoord2Ind, maInd2Coord, maCompInd.

Examples



maCompCoord(1:2,1,1:4,1:3)


[Package marray version 1.18.0 Index]