maCompInd {marray}	R Documentation

Generate spot indices

Description

This function generates spot indices from ranges of rows and columns for the grid and spot matrices. Spots on the array are numbered consecutively starting from the top left grid and the top left spot within each grid.

Usage


maCompInd(grows, gcols, srows, scols, L)

Arguments

grows 	 numeric vector of grid rows.

gcols 	 numeric vector of grid columns.

srows 	 numeric vector of spot rows.

scols 	 numeric vector of spot columns.

L 	 object of class "marrayLayout".

Value

a numeric vector of spot indices.

Author(s)

Yee Hwa (Jean) Yang, Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

See Also

marrayLayout, maCoord2Ind, maInd2Coord, maCompCoord.

Examples


L <- new("marrayLayout", maNgr=4, maNgc=4, maNsr=22, maNsc=24)
maCompInd(1:2,1,1:4,1:3,L)


[Package marray version 1.18.0 Index]