maCompLayout {marray}	R Documentation

Generate a marrayLayout object

Description

Take a matrix of cooordiates and generate a marrayLayout object.

Usage


maCompLayout(mat, ncolumns = 4)

Arguments

mat 	 a matrix of coordinates, this can either be n by 3 matrix with columns (Block, Row, Column) or n by 4 matrix with columns (Grid.R, Grid.C, Spot.R, Spot.C)

ncolumns 	 For n by 3 matrix, the number of meta-grid columns. By default, it is set to 4.

Value

An object of class "marrayLayout".

Author(s)

Jean Yang

Examples


X <- cbind(Block = c(1,1,2,2,3,3,4,4), Rows=c(1,2,1,2,1,2,1,2), Columns=rep(1,8))
maCompLayout(X, ncolumns=2)

[Package marray version 1.18.0 Index]