maCompPlate {marray}	R Documentation

Generate plate IDs

Description

This function generates plate IDs from the dimensions of the grid and spot matrices. Note that this function only applies to arrays with a regular plate layout, where the number of spots is a multiple of the number of wells on a plate (usually 96 or 384) and each well contributes exactly one spot. It should thus be used with caution.

Usage


maCompPlate(x, n=384)

Arguments

x 	 object of class "marrayLayout", "marrayRaw" and "marrayNorm"

n 	 object of class "numeric", number of wells in each plate, usually 384 or 96.

Details

Having plate IDs may be useful for the purpose of normalization. Normalization by plate can be done using the function maNormMain.

Value

a vector of plate IDs (factor).

Author(s)

Yee Hwa (Jean) Yang, Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

See Also

marrayLayout, marrayRaw, marrayNorm

Examples



L<-new("marrayLayout", maNgr=4, maNgc=4, maNsr=22, maNsc=24)
plate<-maCompPlate(L,384)
table(plate)
maPlate(L)<-plate


[Package marray version 1.18.0 Index]