maCoord2Ind {marray}	R Documentation

Convert grid and spot matrix coordinates to spot indices

Description

This functions converts grid and spot matrix coordinates (four coordinates) to spot indices, where spots on the array are numbered consecutively starting from the top left grid and the top left spot within each grid.

Usage


maCoord2Ind(x, L)

Arguments

x 	 a matrix of spot four-coordinates, with rows corresponding to spots and columns to grid row, grid column, spot row, and spot column coordinates.

L 	 an object of class "marrayLayout".

Value

a numeric vector of spot indices.

Author(s)

Yee Hwa (Jean) Yang, Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

See Also

marrayLayout, maInd2Coord, maCompCoord, maCompInd.

Examples


L <- new("marrayLayout", maNgr=4, maNgc=4, maNsr=22, maNsc=24)
coord<-cbind(rep(2,4),rep(1,4),rep(1,4),1:4)
maCoord2Ind(coord, L)


[Package marray version 1.18.0 Index]