maDefaultPar {marray}	R Documentation

Default graphical parameters for microarray objects

Description

This function returns default graphical parameters for microarray objects. The parameters may be passed as arguments to the functions maBoxplot and maPlot.

Usage


maDefaultPar(m, x, y, z)

Arguments

m 	 Microarray object of class "marrayRaw" and "marrayNorm".

x 	 Name of accessor method for the abscissa spot statistic, typically a slot name for the microarray object m, such as maA.

y 	 Name of accessor method for the ordinate spot statistic, typically a slot name for the microarray object m, such as maM.

z 	 Name of accessor method for the spot statistic used to stratify the data, typically a slot name for the microarray layout object (see "marrayLayout") such as maPlate or a method such as maPrintTip.

Value

A list with elements

def.box 	 default graphical parameters for maBoxplot.

def.plot 	 default graphical parameters for maPlot.

def.lines 	 default graphical parameters for functions such as maLoessLines used in maPlot.

def.legend 	 default graphical parameters for functions such as maLegendLines used in maPlot.

def.text 	 default graphical parameters for functions such as maText used in maPlot.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maBoxplot, maPlot, maLegendLines, maLoessLines, maText, maDotsDefaults.

Examples


# See examples for maPlot.


[Package marray version 1.18.0 Index]