maDotsDefaults {marray}	R Documentation

Replace graphical default parameters by user supplied parameters

Description

This function may be used to compare default graphical parameters for microarray diagnostic plots to user supplied parameters given in .... User supplied parameters overwrite the defaults. It is used in maBoxplot, maPlot, and maImage.

Usage


maDotsDefaults(dots, defaults)

Arguments

dots 	 List of user supplied parameters, e.g. from list(...).

defaults 	 List of default parameters, e.g. from the function maDefaultPar.

Value

args 	 List of graphical parameters.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maDefaultPar, maBoxplot, maPlot, maImage.

Examples


dots<-list(xlab="X1", ylab="Y1")
defaults<-list(xlab="X1", ylab="Y2", col=2)
pars<-maDotsDefaults(dots, defaults)

do.call("plot",c(list(x=1:10), pars))


[Package marray version 1.18.0 Index]