maDotsMatch {marray}	R Documentation

Replace default arguments of a function by user supplied values

Description

This function may be used to replace default arguements for any functions to user supplied parameters.

Usage


maDotsMatch(dots, defaults)

Arguments

dots 	 List of user supplied argements, e.g. from list(...).

defaults 	 List of formal arguments of a function, e.g. from the function formals.

Value

args 	 List of argument of a function.

Author(s)

Jean Yee Hwa Yang

See Also

maDefaultPar, maDotsDefaults

Examples


dots<-list(x=1:10, y=11:20)
argsfun <- maDotsMatch(dots, formals(args(plot)))
do.call("plot", argsfun)

[Package marray version 1.18.0 Index]