maGenControls {marray}	R Documentation

Generating a vector recording the control status of the spotted probe sequences.

Description

ControlCode is a matrix representing certain regular expression pattern and the control status of the spotted probe sequences. This function uses `grep' searches for matches to `pattern' (its first argument) within the character vector `x' (second argument).

Usage


maGenControls(Gnames, controlcode, id = "ID")

Arguments

Gnames 	 An object of class matrix, data.frame or marrayInfo which contains description of spotted probe sequences.

controlcode 	 A character matrix of n by 2 columns. The first column contains a few regular expression of spotted probe sequences and the second column contains the corresponding control status.

id 	 the column number of column name in Gnames that contains description of each spot on the array.

Value

A vector of characters recording the control status of the spotted probe sequences.

Author(s)

Jean Yee Hwa Yang

See Also

grep

Examples


data(swirl)
maControls(swirl) <- maGenControls(maGnames(swirl), id="Name")
table(maControls(swirl))

[Package marray version 1.18.0 Index]