maGeneTable {marray}	R Documentation

Table of spot coordinates and gene names

Description

This function produces a table of spot coordinates and gene names for objects of class "marrayRaw" and "marrayNorm".

Usage


maGeneTable(object)

Arguments

object 	 microarray object of class "marrayRaw" and "marrayNorm".

Value

an object of class data.frame, with rows corresponding to spotted probe sequences. The first four columns are the grid matrix and spot matrix coordinates, and the remaining columns are the spot descriptions stored in the maGnames slot of the microarray object.

Author(s)

Yee Hwa (Jean) Yang

See Also

marrayInfo, marrayLayout, marrayRaw, marrayNorm, maCompCoord.

Examples


# Example uses swirl dataset, for description type ? swirl

data(swirl)

tab<-maGeneTable(swirl)
tab[1:10,]


[Package marray version 1.18.0 Index]