maImage.func {marray}	R Documentation

Color image for cDNA microarray spot statistics

Description

This function creates spatial images of shades of gray or colors that correspond to the values of a statistic for each spot on the array. The statistic can be the intensity log-ratio M, a spot quality measure (e.g. spot size or shape), or a test statistic. This function can be used to explore whether there are any spatial effects in the data, for example, print-tip or cover-slip effects. This function is called by maImage.

Usage


maImage.func(x, L, subset=TRUE, col=heat.colors(12), contours=FALSE, overlay=NULL, ol.col=1,  ...)

Arguments

x 	 A "numeric" vector of spot statistics.

L 	 An object of class "marrayLayout", if L is missing we will assume the dimension of x.

subset 	 A "logical" or "numeric" vector indicating the subset of spots to display on the image.

col 	 A list of colors such as that generated by rainbow, heat.colors, topo.colors, terrain.colors, or similar functions. In addition to these color palette functions, a new function maPalette was defined to generate color palettes from user supplied low, middle, and high color values.

contours 	 If contours=TRUE, contours are plotted, otherwise they are not shown.

overlay 	 A logical vector of spots to be highlighted on the image plots.

ol.col 	 Color of the overlay spots.

... 	 Optional graphical parameters, see par.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

image, maImage, maColorBar, maPalette.

Examples


# See examples for image.

[Package marray version 1.18.0 Index]