maInd2Coord {marray}	R Documentation

Convert spot indices to grid and spot matrix coordinates

Description

This functions converts spot indices to grid and spot matrix coordinates (four coordinates), where spots on the array are numbered consecutively starting from the top left grid and the top left spot within each grid.

Usage


maInd2Coord(x, L)

Arguments

x 	 a numeric vector of spot indices.

L 	 an object of class "marrayLayout".

Value

a matrix of spot four-coordinates, with rows corresponding to spots and columns to grid row, grid column, spot row, and spot column coordinates.

Author(s)

Yee Hwa (Jean) Yang, Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

See Also

marrayLayout, maCoord2Ind, maCompCoord, maCompInd.

Examples



L <- new("marrayLayout", maNgr=4, maNgc=4, maNsr=22, maNsc=24)
maInd2Coord(c(1:10,529:538), L)


[Package marray version 1.18.0 Index]