maLoess {marray}	R Documentation

Stratified univariate robust local regression

Description

This function performs robust local regression of a variable y on predictor variable x, separately within values of a third variable z. It is used by maNormLoess for intensity dependent location normalization.

Usage


maLoess(x, y, z, w=NULL, subset=TRUE, span=0.4, ...)

Arguments

x 	 A numeric vector of predictor variables.

y 	 A numeric vector of responses.

z 	 Variables used to stratify the data.

w 	 An optional numeric vector of weights.

subset 	 A "logical" or "numeric" vector indicating the subset of points used to compute the fits.

span 	 The argument span which controls the degree of smoothing in the loess function.

... 	 Misc arguments.

Details

y is regressed on x, separately within values of z using the loess function.

Value

A numeric vector of fitted values.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

Y. H. Yang, S. Dudoit, P. Luu, and T. P. Speed (2001). Normalization for cDNA microarray data. In M. L. Bittner, Y. Chen, A. N. Dorsel, and E. R. Dougherty (eds), Microarrays: Optical Technologies and Informatics, Vol. 4266 of Proceedings of SPIE.

Y. H. Yang, S. Dudoit, P. Luu, D. M. Lin, V. Peng, J. Ngai, and T. P. Speed (2002). Normalization for cDNA microarray data: a robust composite method addressing single and multiple slide systematic variation. Nucleic Acids Research, Vol. 30, No. 4.

See Also

maNormMain, maNormLoess, loess.

Examples


# See examples for maNormMain.

[Package marray version 1.18.0 Index]