maLoessLines {marray}	R Documentation

Add smoothed fits to a plot

Description

This function may be used to compute and plot loess or lowess fits for an existing plot. The plot can be produced by plot, maPlot, or maPlot.func.

Usage


maLoessLines(subset=TRUE, weights=NULL, loess.args=list(span = 0.4,
degree=1, family="symmetric", control=loess.control(trace.hat =
"approximate", iterations=5, surface="direct")), col=2, lty=1, lwd=2.5, ...)

maLowessLines(subset = TRUE, f = 0.3, col = 2, lty = 1, lwd = 2.5, ...)

Arguments

subset 	 A "logical" or "numeric" vector indicating the subset of points used to compute the fits.

weights 	 Optional "numeric" vector of weights – for maLoessLines only.

loess.args 	 List of optional arguments for the loess functions – for maLoessLines only.

f 	 The smoother span for the lowess function – for maLowessLines only.

col 	 The fitted line colors.

lty 	 The fitted line types.

lwd 	 The fitted line widths.

... 	 Optional graphical parameters, see par.

Value

A function with bindings for subset, weights, loess.args, col, lty, lwd, and .... This latter function takes as arguments x and y, the abscissa and ordinates of points on the plot, and z a vector of discrete values used to stratify the points. Loess (or lowess) fits are performed separately within values of z.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

loess, lowess, maPlot, maPlot.func.

Examples


# See examples for maPlot.


[Package marray version 1.18.0 Index]