maMAD {marray}	R Documentation

Stratified MAD calculation

Description

This function computes the median absolute deviation (MAD) of values in y separately within values of x. It is used by maNormMAD for MAD scale normalization.

Usage


maMAD(x, y, geo=TRUE, subset=TRUE)

Arguments

x 	 Variables used to stratify the data.

y 	 A numeric vector.

geo 	 If TRUE, the MAD of each group is divided by the geometric mean of the MADs across groups (cf. Yang et al. (2002)). This allows observations to retain their original units.

subset 	 A "logical" or "numeric" vector indicating the subset of points used to compute the MAD.

Value

A numeric vector of MAD values.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

Y. H. Yang, S. Dudoit, P. Luu, and T. P. Speed (2001). Normalization for cDNA microarray data. In M. L. Bittner, Y. Chen, A. N. Dorsel, and E. R. Dougherty (eds), Microarrays: Optical Technologies and Informatics, Vol. 4266 of Proceedings of SPIE.

Y. H. Yang, S. Dudoit, P. Luu, D. M. Lin, V. Peng, J. Ngai, and T. P. Speed (2002). Normalization for cDNA microarray data: a robust composite method addressing single and multiple slide systematic variation. Nucleic Acids Research, Vol. 30, No. 4.

See Also

maNormMain, maNormMAD, mad.

Examples


# See examples for maNormMain.


[Package marray version 1.18.0 Index]