na {marray}	R Documentation

Basic Statistical Functions for Handling Missing Values

Description

Basic statistical functions for handling missing values or NA.
In log.na, sum.na, mean.na and var.na, quantile.na, length.na, missing values are omitted from the calculation.
The function cor.na calls cor with the argument use="pairwise.complete.obs".
The function order.na only handles vector arguments and not lists. However, it gives the option of omitting the NAs (na.last=NA), of placing the NAs at the start of the ordered vector (na.last=F) or at the end (na.last=T).
The function scale.na is a modified version of scale which allows NAs in the variance calculation. If scale = T, the function f in scale.na uses var.na to perform the variance calculation. The function prod.na is similar to the prod function with na.rm=TRUE. This function returns the product of all the values present in its arguments, omitting any missing values.

Author(s)

Yee Hwa Yang, jean@biostat.berkeley.edu

See Also

log, sum, mean, var, cor, order, scale, prod.

[Package marray version 1.18.0 Index]