maNorm {marray}	R Documentation

Simple location and scale normalization function

Description

This function is a simple wrapper function around the main normalization function maNormMain. It allows the user to choose from a set of six basic location and scale normalization procedures. The function operates on an object of class "marrayRaw" (or possibly "marrayNorm", if normalization is performed in several steps) and returns an object of class "marrayNorm".

Usage


maNorm(mbatch, norm=c("printTipLoess", "none", "median", "loess",
"twoD", "scalePrintTipMAD"), subset=TRUE, span=0.4, Mloc=TRUE,
Mscale=TRUE, echo=FALSE, ...)

Arguments

mbatch 	 Object of class marrayRaw, containing intensity data for the batch of arrays to be normalized. An object of class "marrayNorm" may also be passed if normalization is performed in several steps.

norm 	 Character string specifying the normalization procedures:

none

no normalization

median

for global median location normalization

loess

for global intensity or A-dependent location normalization using the loess function

twoD

for 2D spatial location normalization using the loess function

printTipLoess

for within-print-tip-group intensity dependent location normalization using the loess function

scalePrintTipMAD

for within-print-tip-group intensity dependent location normalization followed by within-print-tip-group scale normalization using the median absolute deviation (MAD).

This argument can be specified using the first letter of each method.

subset 	 A "logical" or "numeric" vector indicating the subset of points used to compute the normalization values.

span 	 The argument span which controls the degree of smoothing in the loess function.

Mloc 	 If TRUE, the location normalization values are stored in the slot maMloc of the object of class "marrayNorm" returned by the function, if FALSE, these values are not retained.

Mscale 	 If TRUE, the scale normalization values are stored in the slot maMscale of the object of class "marrayNorm" returned by the function, if FALSE, these values are not retained.

echo 	 If TRUE, the index of the array currently being normalized is printed.

... 	 Misc arguments

Details

See maNormMain for details and also more general procedures.

Value

mnorm 	 An object of class "marrayNorm", containing the normalized intensity data.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

Y. H. Yang, S. Dudoit, P. Luu, and T. P. Speed (2001). Normalization for cDNA microarray data. In M. L. Bittner, Y. Chen, A. N. Dorsel, and E. R. Dougherty (eds), Microarrays: Optical Technologies and Informatics, Vol. 4266 of Proceedings of SPIE.

Y. H. Yang, S. Dudoit, P. Luu, D. M. Lin, V. Peng, J. Ngai, and T. P. Speed (2002). Normalization for cDNA microarray data: a robust composite method addressing single and multiple slide systematic variation. Nucleic Acids Research, Vol. 30, No. 4.

See Also

maNormMain, maNormScale.

Examples


# Examples use swirl dataset, for description type ? swirl
data(swirl)

# Global median normalization for swirl arrays 2 and 3
mnorm<-maNorm(swirl[,2:3], norm="median", echo=TRUE)

# Within-print-tip-group loess location normalization for swirl array 1
mnorm<-maNorm(swirl[,1], norm="p", span=0.45)


[Package marray version 1.18.0 Index]