maNormLoess {marray}	R Documentation

Intensity dependent location normalization function

Description

This function is used for intensity dependent location normalization, using the robust local regression function loess. It should be used as an argument to the main normalization function maNormMain.

Usage


maNormLoess(x="maA", y="maM", z="maPrintTip", w=NULL, subset=TRUE,
span=0.4, ...)

Arguments

x 	 Name of accessor method for spot statistics, usually maA.

y 	 Name of accessor method for spot statistics, usually maM.

z 	 Name of accessor method for spot statistic used to stratify the data, usually a layout parameter, e.g. maPrintTip or maPlate. If z is not a character, e.g. NULL, the data are not stratified.

w 	 An optional numeric vector of weights.

subset 	 A "logical" or "numeric" vector indicating the subset of points used to compute the fits.

span 	 The argument span which controls the degree of smoothing in the loess function.

... 	 Misc arguments

Value

A function with bindings for the above arguments. This latter function takes as argument an object of class "marrayRaw" (or possibly "marrayNorm"), and returns a vector of fitted values to be substracted from the raw log-ratios. It calls the function maLoess, which is not specific to microarray objects.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

Y. H. Yang, S. Dudoit, P. Luu, and T. P. Speed (2001). Normalization for cDNA microarray data. In M. L. Bittner, Y. Chen, A. N. Dorsel, and E. R. Dougherty (eds), Microarrays: Optical Technologies and Informatics, Vol. 4266 of Proceedings of SPIE.

Y. H. Yang, S. Dudoit, P. Luu, D. M. Lin, V. Peng, J. Ngai, and T. P. Speed (2002). Normalization for cDNA microarray data: a robust composite method addressing single and multiple slide systematic variation. Nucleic Acids Research, Vol. 30, No. 4.

See Also

maNormMain, maLoess, loess.

Examples


# See examples for maNormMain.


[Package marray version 1.18.0 Index]