maNum2Logic {marray}	R Documentation

Convert a numeric vector of indices to a logical vector

Description

This function converts a numeric vector of indices to a logical vector. It is used for subsetting purposes.

Usage


maNum2Logic(n=length(subset), subset=TRUE)

Arguments

n 	 the length of the logical vector to be produced.

subset 	 a numeric vector of indices. A logical vector may also be supplied, in which case it is also the value of the function.

Value

a logical vector.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

Examples


maNum2Logic(10, 1:3)

[Package marray version 1.18.0 Index]