maPlot {marray}	R Documentation

Scatter-plots for cDNA microarray spot statistics

Description

The function maPlot produces scatter-plots of microarray spot statistics for the classes "marrayRaw" and "marrayNorm". It also allows the user to highlight and annotate subsets of points on the plot, and display fitted curves from robust local regression or other smoothing procedures.

Usage


maPlot(m, x="maA", y="maM", z="maPrintTip", lines.func, text.func, legend.func, ...)

Arguments

m 	 Microarray object of class "marrayRaw" and "marrayNorm".

x 	 Name of accessor function for the abscissa spot statistic, typically a slot name for the microarray object m, such as maA.

y 	 Name of accessor function for the ordinate spot statistic, typically a slot name for the microarray object m, such as maM.

z 	 Name of accessor method for the spot statistic used to stratify the data, typically a slot name for the microarray layout object (see "marrayLayout") such as maPlate or a method such as maPrintTip. If z is NULL, the data are not stratified.

lines.func 	 Function for computing and plotting smoothed fits of y as a function of x, separately within values of z, e.g. maLoessLines. If lines.func is NULL, no fitting is performed.

text.func 	 Function for highlighting a subset of points, e.g., maText. If text.func is NULL, no points are highlighted.

legend.func 	 Function for adding a legend to the plot, e.g. maLegendLines. If legend.func is NULL, there is no legend.

... 	 Optional graphical parameters, see par.

Details

This function calls the general function maPlot.func, which is not specific to microarray data. If there are more than one array in the batch, the plot is done for the first array, by default. Default graphical parameters are chosen for convenience using the function maDefaultPar (e.g. color palette, axis labels, plot title) but the user has the option to overwrite these parameters at any point.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maPlot.func, maDefaultPar, maLoessLines, maLegendLines, maText, plot, lowess, loess, legend.

Examples


# To see the demo type demo(marrayPlots)

# Examples use swirl dataset, for description type ? swirl
data(swirl)

# - Default arguments
maPlot(swirl)

# Lowess fit using all spots
maPlot(swirl, z=NULL, legend.func=NULL)

# Loess fit using all spots
maPlot(swirl, z=NULL, legend.func=maLegendLines(legend="All spots",col="green"), lines.func=maLoessLines(loess.args=list(span=0.3),col="green"))

# Pre-normalization MA-plot for the Swirl 81 array, with the lowess fits for 
# individual grid columns and 1% tails of M highlighted
defs <- maDefaultPar(swirl[, 1], x = "maA", y = "maM", z = "maGridCol")
legend.func <- do.call("maLegendLines", defs$def.legend)
lines.func <- do.call("maLowessLines", c(list(TRUE, f = 0.3), defs$def.lines))
text.func<-maText(subset=maTop(maM(swirl)[,1],h=0.01,l=0.01), labels="o", col="violet")
maPlot(swirl[, 1], x = "maA", y = "maM", z = "maGridCol", lines.func=lines.func, text.func = text.func, legend.func=legend.func, main = "Swirl array 81: pre-normalization MA-plot")

[Package marray version 1.18.0 Index]