maPlot.func {marray}	R Documentation

Scatter-plots with fitted curves and text

Description

This function produces scatter-plots of x vs. y. It also allows the user to highlight and annotate subsets of points on the plot, and display fitted curves from robust local regression or other smoothing procedures.

Usage


maPlot.func(x, y, z, 
        lines.func = maLowessLines(subset = TRUE, f = 0.3, col = 1:length(unique(z)), lty = 1, lwd = 2.5),
        text.func = maText(), 
        legend.func = maLegendLines(legend = as.character(unique(z)), col = 1:length(unique(z)), lty = 1, lwd = 2.5, ncol = 1),
         ...)

Arguments

x 	 A "numeric" vector for the abscissa.

y 	 A "numeric" vector for the ordinates.

z 	 A vector of statistic used to stratify the data, smoothed curves are fitted separately within values of z

lines.func 	 A function for computing and plotting smoothed fits of y as a function of x, separately within values of z, e.g. maLoessLines.

text.func 	 A function for highlighting a subset of points, e.g., maText.

legend.func 	 A function for adding a legend to the plot, e.g. maLegendLines.

... 	 Optional graphical parameters, see par.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maPlot, maLoessLines, maLegendLines, maText, plot, lowess, loess, legend.

Examples


# See examples for maPlot.


[Package marray version 1.18.0 Index]