maText {marray}	R Documentation

Highlight points on a plot

Description

This function may be used to highlight a subset of points on an existing plot, such as a plot produced by plot, maPlot, or maPlot.func.

Usage


maText(subset=NULL, labels=as.character(1:length(subset)), ...)

Arguments

subset 	 A "logical" or "numeric" vector indicating the subset of points to highlight.

labels 	 One or more character strings or expressions specifying the text to be written.

... 	 Optional graphical parameters, see par.

Value

A function with bindings for subset, labels, and .... This latter function takes as arguments x and y, the absissa and ordinates of points on the plot.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

text, maPlot, maPlot.func.

Examples


# See examples for maPlot.


[Package marray version 1.18.0 Index]