maTop {marray}	R Documentation

Identify extreme values

Description

This function determines which values in a numeric vector are above or below user supplied cut-offs.

Usage


maTop(x, h=1, l=1)

Arguments

x 	 A "numeric" vector.

h 	 A "numeric", upper cut-off.

l 	 A "numeric", lower cut-off.

Value

A "logical" vector indicating which entries are above or below the cut-offs.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maPlot, maImage, quantile.

Examples


# See examples for maPlot.

[Package marray version 1.18.0 Index]