maTwoSamples {marray}	R Documentation

Changing signs for two sample analysis

Description

Taking target file information and flip the dye swaps experiments.

Usage


maTwoSamples(targetfile, normdata, Trt, Ctl, targetID = "TargetName", slidesID = "Slides", dyesID = "Dyes", RedID = 5, path = ".", output = TRUE)

Arguments

targetfile 	 A data.frame containing target samples information.

normdata 	 A R object of class 'marrayNorm'

Trt 	 A character string representing "treatment" sample.

Ctl 	 A character string representing "controls" sample.

targetID 	 A character string representing the column name in 'targetfile' containing target samples information.

slidesID 	 A character string representing the column name in 'targetfile' containing the slide label.

dyesID 	 A character string representing the column name in 'targetfile' containing dye labeled information.

RedID 	 The character use to represent the Cy5 dye.

path 	 A character string representing the data directory. By default this is set to the current working directory (".").

output 	 Save and tab delimited file

Value

An objects of 'marrayNorm' with the dye assignment adjusted.

Author(s)

Yee Hwa (Jean) Yang

[Package marray version 1.18.0 Index]