mapGeneInfo {marray}	R Documentation

Creating URL strings for external database links

Description

These functions are used with htmlPage. The function mapGeneInfo, takes all the arguments and generate a character matrix of two columns. The first columns representing the name of the argument and the second columns represents the value of an argument. The function widget.mapGeneInfo allows the user to enter this information interactively.

Usage


mapGeneInfo(widget = FALSE, Gnames, Name = "pubmed", ID =
                 "genbank", ACC = "SMDacc", ...)
widget.mapGeneInfo(Gnames) 

Arguments

widget 	 A logical value specifying if widgets should be used.

Name 	 The external database for spot description, E.g. "pubmed".

ID 	 The external database for spot ID, E.g. "operon", "Riken", "locuslink".

ACC 	 The external database for gene accession number, E.g. "genebank".

Gnames 	 An object of class matrix, data.frame or marrayInfo which contains description of spotted probe sequences.

... 	 Other column names

Details

The function mapGeneInfo generates a character matrix with the first column representing the column headings of "Gnames" and the second column representing the corresponding names in the list URLstring. For example, if a particular column in "Gnames" with column names "ID" contains genebank accession number, then the function mapGeneInfo generates a row containing "ID" in the first column and "genbank" in the second. Examples are SFGL and UCBFGL.

URLstring is a list contains the URL to various external database, E.g. operon, Riken, genbank.
The current choices are: "pubmed", "locuslink", "riken", "SMDclid", "SMDacc", "operonh2", "operonh1" , "operonm2", "operonm1" and "genbank" . "SMDclid" and "SMDacc" are links to Stanford Microarray Databases.

Author(s)

Jean Yee Hwa Yang

Examples


 mapGeneInfo(ID="genebank", ll="locuslink")
 mapGeneInfo(ID="locuslink", Sample.ID="riken")

[Package marray version 1.18.0 Index]