marrayInfo-class {marray}	R Documentation

Class "marrayInfo", description of target samples or spotted probe sequences

Description

This class is used to store information on target samples hybridized to a batch of arrays or probe sequences spotted onto these arrays. It is not specific to the microarray context.

Objects from the Class

Objects can be created by calls of the form new('marrayInfo',
maLabels = ...., # Object of class character
maInfo = ...., # Object of class data.frame
maNotes = ...., # Object of class character
)

Slots

maLabels:

Object of class "character", vector of spot or array labels.

maInfo:

Object of class "data.frame". If the object of class "marrayInfo" is used to describe probe sequences, rows of maInfo correspond to spots and columns to various gene identifiers and annotations. If the object of class "marrayInfo" is used to describe target samples hybridized to the arrays, rows of maInfo correspond to arrays and columns to various descriptions of the hybridizations, e.g., names of Cy3 and Cy5 samples, labels for the arrays etc.

maNotes:

Object of class "character", any notes on the target samples or spotted probe sequences.

Methods

[

signature(x = "marrayInfo"): subsetting operator for spots on the array or arrays in the batch, ensures that all slots are subset properly.

maGnames<-

signature(object = "marrayRaw", value = "marrayInfo"): slot assignment method.

maGnames<-

signature(object = "marrayNorm", value = "marrayInfo"): slot assignment method.

maGnames<-

signature(object = "marraySpots", value = "marrayInfo"): slot assignment method.

maInfo

signature(object = "marrayInfo"): slot accessor method.

maInfo<-

signature(object = "marrayInfo", value = "data.frame"): slot assignment method.

maLabels

signature(object = "marrayInfo"): slot accessor method.

maLabels<-

signature(object = "marrayInfo", value = "character"): slot assignment method.

maLabels<-

signature(object = "marrayInfo", value = "numeric"): slot assignment method.

maNotes

signature(object = "marrayInfo"): slot accessor method.

maNotes<-

signature(object = "marrayInfo", value = "character"): slot assignment method.

maTargets<-

signature(object = "marrayRaw", value = "marrayInfo"): slot assignment method.

maTargets<-

signature(object = "marrayNorm", value = "marrayInfo"): slot assignment method.

print

signature(x = "marrayInfo"): print method for "marrayInfo" class.

Author(s)

Jean Yang and Sandrine Dudoit

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

marrayLayout, marrayRaw, marrayNorm.

Examples


## See marrayRaw

[Package marray version 1.18.0 Index]