marrayLayout-class {marray}	R Documentation

Class "marrayLayout", classes and methods for layout parameters of cDNA microarrays

Description

This class is used to keep track of important layout parameters for two-color cDNA microarrays. It contains slots for: the total number of spotted probe sequences on the array, the dimensions of the spot and grid matrices, the plate origin of the probes, information on spotted control sequences (e.g. probe sequences which should have equal abundance in the two target samples, such as housekeeping genes). The terms print-tip-group, grid, spot matrix, and sector are used interchangeably and refer to a set of spots printed using the same print-tip.

Objects from the Class

Objects can be created by calls of the form new('marrayLayout',
maNgr = ...., # Object of class numeric
maNgc = ...., # Object of class numeric
maNsr = ...., # Object of class numeric
maNsc = ...., # Object of class numeric
maNspots = ...., # Object of class numeric
maSub = ...., # Object of class logical
maPlate = ...., # Object of class factor
maControls = ...., # Object of class factor
maNotes = ...., # Object of class character
)

Slots

maNgr:

Object of class "numeric", number of rows for the grid matrix.

maNgc:

Object of class "numeric", number of columns for the grid matrix.

maNsr:

Object of class "numeric", number of rows for the spot matrices.

maNsc:

Object of class "numeric", number of columns for the spot matrices.

maNspots:

Object of class "numeric", total number of spots on the array, equal to maNgr x maNgc x maNsr x maNsc.

maSub:

Object of class "logical", indicating which spots are currently being considered.

maPlate:

Object of class "factor", recording the plate origin of the spotted probe sequences.

maControls:

Object of class "factor", recording the control status of the spotted probe sequences.

maNotes:

Object of class "character", any notes concerning the microarray layout, e.g., printing conditions.

Methods

[

signature(x = "marrayLayout"): subsetting operator for spots on the array, ensures that all slots are subset properly.

maControls<-

signature(object = "marrayLayout"): slot assignment method.

maControls

signature(object = "marrayLayout"): slot accessor method.

maGridCol

signature(object = "marrayLayout"): method which computes a vector of grid column coordinates for each spot.

maGridRow

signature(object = "marrayLayout"): method which computes a vector of grid row coordinates for each spot.

maLayout<-

signature(object = "marrayRaw", value = "marrayLayout"): slot assignment method.

maLayout<-

signature(object = "marrayNorm", value = "marrayLayout"): slot assignment method.

maNgc

signature(object = "marrayLayout"): slot accessor method.

maNgc<-

signature(object = "marrayLayout", value = "numeric"): slot assignment method.

maNgr

signature(object = "marrayLayout"): slot accessor method.

maNgr<-

signature(object = "marrayLayout", value = "numeric"): slot assignment method.

maNotes

signature(object = "marrayLayout"): slot accessor method.

maNotes<-

signature(object = "marrayLayout", value = "character"): slot assignment method.

maNsc

signature(object = "marrayLayout"): slot accessor method.

maNsc<-

signature(object = "marrayLayout", value = "numeric"): slot assignment method.

maNspots

signature(object = "marrayLayout"): slot accessor method.

maNspots<-

signature(object = "marrayLayout", value = "numeric"): slot assignment method.

maNsr

signature(object = "marrayLayout"): slot accessor method.

maNsr<-

signature(object = "marrayLayout", value = "numeric"): slot assignment method.

maPlate

signature(object = "marrayLayout"): slot accessor method.

maPlate<-

signature(object = "marrayLayout"): slot assignment method.

maPrintTip

signature(object = "marrayLayout"): method which computes a vector of print-tip-group indices for each spot.

maSpotCol

signature(object = "marrayLayout"): method which computes a vector of spot column coordinates for each spot.

maSpotRow

signature(object = "marrayLayout"): method which computes a vector of spot row coordinates for each spot.

maSub

signature(object = "marrayLayout"): slot accessor method.

maSub<-

signature(object = "marrayLayout", value = "logical"): slot assignment method.

maSub<-

signature(object = "marrayLayout", value = "numeric"): slot assignment method.

print

signature(x = "marrayLayout"): print method for "marrayLayout" class.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

marrayRaw, marrayNorm, marrayInfo and [-methods.

Examples


## See marrayRaw

[Package marray version 1.18.0 Index]