marrayNorm-class {marray}	R Documentation

Class "marrayNorm", classes and methods for post-normalization cDNA microarray intensity data

Description

This class represents post-normalization intensity data for a batch of cDNA microarrays. A batch of arrays consists of a collection of arrays with the same layout ("marrayLayout"). The class contains slots for the average log-intensities A, the normalized log-ratios M, the location and scale normalization values, the layout of the arrays, and descriptions of the target samples hybridized to the arrays and probe sequences spotted onto the arrays.

Objects from the Class

Objects can be created by calls of the form new('marrayNorm',
maA = ...., # Object of class matrix
maM = ...., # Object of class matrix
maMloc = ...., # Object of class matrix
maMscale = ...., # Object of class matrix
maW = ...., # Object of class matrix
maLayout = ...., # Object of class marrayLayout
maGnames = ...., # Object of class marrayInfo
maTargets = ...., # Object of class marrayInfo
maNotes = ...., # Object of class character
maNormCall = ...., # Object of class call
)

Slots

maA:

Object of class "matrix", average log-intensities (base 2) A, rows correspond to spotted probe sequences, columns to arrays in the batch.

maM:

Object of class "matrix", intensity log-ratios (base 2) M, rows correspond to spotted probe sequences, columns to arrays in the batch.

maMloc:

Object of class "matrix", location normalization values, rows correspond to spotted probe sequences, columns to arrays in the batch.

maMscale:

Object of class "matrix", scale normalization values, rows correspond to spotted probe sequences, columns to arrays in the batch.

maW:

Object of class "matrix", spot quality weights, rows correspond to spotted probe sequences, columns to arrays in the batch.

maLayout:

Object of class "marrayLayout", layout parameters for cDNA microarrays.

maGnames:

Object of class "marrayInfo", description of spotted probe sequences.

maTargets:

Object of class "marrayInfo", description of target samples hybridized to the arrays.

maNotes:

Object of class "character", any notes concerning the microarray experiments, e.g. hybridization or scanning conditions.

maNormCall:

Object of class "call", function call for normalizing the batch of arrays.

Methods

[

signature(x = "marrayNorm"): subsetting operator for spots on the array and arrays in the batch, ensures that all slots are subset properly.

coerce

signature(from = "marrayRaw", to = "marrayNorm"): coerce an object of class "marrayRaw" into an object of class marrayNorm.

maA

signature(object = "marrayNorm"): slot accessor method.

maA<-

signature(object = "marrayNorm", value = "matrix"): slot assignment method.

maControls<-

signature(object = "marrayNorm"): slot assignment method.

maControls

signature(object = "marrayNorm"): slot accessor method.

maGnames

signature(object = "marrayNorm"): slot accessor method.

maGnames<-

signature(object = "marrayNorm", value = "marrayInfo"): slot assignment method.

maGridCol

signature(object = "marrayNorm"): method which computes a vector of grid column coordinates for each spot.

maGridRow

signature(object = "marrayNorm"): method which computes a vector of grid row coordinates for each spot.

maLayout

signature(object = "marrayNorm"): slot accessor method.

maLayout<-

signature(object = "marrayNorm", value = "marrayLayout"): slot assignment method.

maM

signature(object = "marrayNorm"): slot accessor method.

maM<-

signature(object = "marrayNorm", value = "matrix"): slot assignment method.

maMloc

signature(object = "marrayNorm"): slot accessor method.

maMloc<-

signature(object = "marrayNorm", value = "matrix"): slot assignment method.

maMscale

signature(object = "marrayNorm"): slot accessor method.

maMscale<-

signature(object = "marrayNorm", value = "matrix"): slot assignment method.

maNgc

signature(object = "marrayNorm"): slot accessor method.

maNgc<-

signature(object = "marrayNorm", value = "numeric"): slot assignment method.

maNgr

signature(object = "marrayNorm"): slot accessor method.

maNgr<-

signature(object = "marrayNorm", value = "numeric"): slot assignment method.

maNormCall

signature(object = "marrayNorm"): slot accessor method.

maNotes

signature(object = "marrayNorm"): slot accessor method.

maNotes<-

signature(object = "marrayNorm", value = "character"): slot assignment method.

maNsamples

signature(object = "marrayNorm"): slot accessor method.

maNsc

signature(object = "marrayNorm"): slot accessor method.

maNsc<-

signature(object = "marrayNorm", value = "numeric"): slot assignment method.

maNspots

signature(object = "marrayNorm"): slot accessor method.

maNspots<-

signature(object = "marrayNorm", value = "numeric"): slot assignment method.

maNsr

signature(object = "marrayNorm"): slot accessor method.

maNsr<-

signature(object = "marrayNorm", value = "numeric"): slot assignment method.

maPlate

signature(object = "marrayNorm"): slot accessor method.

maPlate<-

signature(object = "marrayNorm"): slot assignment method.

maPrintTip

signature(object = "marrayNorm"): method which computes a vector of print-tip-group indices for each spot.

maSpotCol

signature(object = "marrayNorm"): method which computes a vector of spot column coordinates for each spot.

maSpotRow

signature(object = "marrayNorm"): method which computes a vector of spot row coordinates for each spot.

maSub

signature(object = "marrayNorm"): slot accessor method.

maSub<-

signature(object = "marrayNorm"): slot assignment method.

maTargets

signature(object = "marrayNorm"): slot accessor method.

maTargets<-

signature(object = "marrayNorm", value = "marrayInfo"): slot assignment method.

maW

signature(object = "marrayNorm"): slot accessor method.

maW<-

signature(object = "marrayNorm", value = "matrix"): slot assignment method.

print

signature(x = "marrayNorm"): print method for "marrayNorm" class.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

marrayLayout, marrayRaw, marrayInfo

Examples


# Examples use swirl dataset, for description type ? swirl

data(swirl)

# Median normalization
mnorm<-maNorm(swirl[,2:3],norm="m")

# Object of class marrayNorm for the second and third swirl arrays
mnorm

# Function call
maNormCall(mnorm)

#  Object of class marrayInfo -- Probe sequences
maGnames(mnorm)

#  Object of class marrayInfo -- Target samples
 maTargets(mnorm)

# Density plot of log-ratios M for third array
plot(density(maM(mnorm[,2])), lwd=2, col=2, main="Density plots of log-ratios M")
lines(density(maM(swirl[,3])), lwd=2)
abline(v=0)
legend(2,1,c("Pre-normalization","Post-normalization"))

[Package marray version 1.18.0 Index]