marrayRaw-class {marray}	R Documentation

Class "marrayRaw", classes and methods for pre-normalization cDNA microarray intensity data

Description

This class represents pre-normalization intensity data for a batch of cDNA microarrays. A batch of arrays consists of a collection of arrays with the same layout ("marrayLayout"). The class contains slots for the green (Cy3) and red (Cy5) foreground and background intensities, the layout of the arrays, and descriptions of the target samples hybridized to the arrays and probe sequences spotted onto the arrays.

Objects from the Class

Objects can be created by calls of the form new('marrayRaw',
maRf = ...., # Object of class matrix
maGf = ...., # Object of class matrix
maRb = ...., # Object of class matrix
maGb = ...., # Object of class matrix
maW = ...., # Object of class matrix
maLayout = ...., # Object of class marrayLayout
maGnames = ...., # Object of class marrayInfo
maTargets = ...., # Object of class marrayInfo
maNotes = ...., # Object of class character
)

Slots

maRf:

Object of class "matrix", red foreground intensities, rows correspond to spotted probe sequences, columns to arrays in the batch.

maGf:

Object of class "matrix", green foreground intensities, rows correspond to spotted probe sequences, columns to arrays in the batch.

maRb:

Object of class "matrix", red background intensities, rows correspond to spotted probe sequences, columns to arrays in the batch.

maGb:

Object of class "matrix", green background intensities, rows correspond to spotted probe sequences, columns to arrays in the batch.

maW:

Object of class "matrix", spot quality weights, rows correspond to spotted probe sequences, columns to arrays in the batch.

maLayout:

Object of class "marrayLayout", layout parameters for the cDNA microarrays.

maGnames:

Object of class "marrayInfo", description of spotted probe sequences.

maTargets:

Object of class "marrayInfo", description of target samples hybridized to the arrays.

maNotes:

Object of class "character", any notes concerning the microarray experiments, e.g. hybridization or scanning conditions.

Methods

[

signature(x = "marrayRaw"): subsetting operator for spots on the array and arrays in the batch, ensures that all slots are subset properly.

coerce

signature(from = "marrayRaw", to = "marrayNorm"): coerce an object of class "marrayRaw" into an object of class "marrayNorm".

maA

signature(object = "marrayRaw"): function which computes average log-intensities (base 2) A for an object of class "marrayRaw".

maControls<-

signature(object = "marrayRaw"): slot assignment method.

maControls

signature(object = "marrayRaw"): slot accessor method.

maGb

signature(object = "marrayRaw"): slot accessor method.

maGb<-

signature(object = "marrayRaw", value = "matrix"): slot assignment method.

maGb<-

signature(object = "marrayRaw", value = "NULL"): slot assignment method.

maGf

signature(object = "marrayRaw"): slot accessor method.

maGf<-

signature(object = "marrayRaw", value = "matrix"): slot assignment method.

maGnames

signature(object = "marrayRaw"): slot accessor method.

maGnames<-

signature(object = "marrayRaw", value = "marrayInfo"): slot assignment method.

maGridCol

signature(object = "marrayRaw"): method which computes a vector of grid column coordinates for each spot.

maGridRow

signature(object = "marrayRaw"): method which computes a vector of grid row coordinates for each spot.

maLayout

signature(object = "marrayRaw"): slot accessor method.

maLayout<-

signature(object = "marrayRaw", value = "marrayLayout"): slot assignment method.

maLG

signature(object = "marrayRaw"): method which computes green log-intensities (base 2) for an object of class "marrayRaw".

maLR

signature(object = "marrayRaw"): method which computes red log-intensities (base 2) for an object of class "marrayRaw".

maM

signature(object = "marrayRaw"): method which computes intensity log-ratios (base 2) M for an object of class "marrayRaw".

maNgc

signature(object = "marrayRaw"): slot accessor method.

maNgc<-

signature(object = "marrayRaw", value = "numeric"): slot assignment method.

maNgr

signature(object = "marrayRaw"): slot accessor method.

maNgr<-

signature(object = "marrayRaw", value = "numeric"): slot assignment method.

maNotes

signature(object = "marrayRaw"): slot accessor method.

maNotes<-

signature(object = "marrayRaw", value = "character"): slot assignment method.

maNsamples

signature(object = "marrayRaw"): slot accessor method.

maNsc

signature(object = "marrayRaw"): slot accessor method.

maNsc<-

signature(object = "marrayRaw", value = "numeric"): slot assignment method.

maNspots

signature(object = "marrayRaw"): slot accessor method.

maNspots<-

signature(object = "marrayRaw", value = "numeric"): slot assignment method.

maNsr

signature(object = "marrayRaw"): slot accessor method.

maNsr<-

signature(object = "marrayRaw", value = "numeric"): slot assignment method.

maPlate

signature(object = "marrayRaw"): slot accessor method.

maPlate<-

signature(object = "marrayRaw"): slot assignment method.

maPrintTip

signature(object = "marrayRaw"): method which computes a vector of print-tip-group indices for each spot.

maRb

signature(object = "marrayRaw"): slot accessor method.

maRb<-

signature(object = "marrayRaw", value = "matrix"): slot assignment method.

maRb<-

signature(object = "marrayRaw", value = "NULL"): slot assignment method.

maRf

signature(object = "marrayRaw"): slot accessor method.

maRf<-

signature(object = "marrayRaw", value = "matrix"): slot assignment method.

maSpotCol

signature(object = "marrayRaw"): method which computes a vector of spot column coordinates for each spot.

maSpotRow

signature(object = "marrayRaw"): method which computes a vector of spot row coordinates for each spot.

maSub

signature(object = "marrayRaw"): slot accessor method.

maSub<-

signature(object = "marrayRaw"): slot assignment method.

maTargets

signature(object = "marrayRaw"): slot accessor method.

maTargets<-

signature(object = "marrayRaw", value = "marrayInfo"): slot assignment method.

maW

signature(object = "marrayRaw"): slot accessor method.

maW<-

signature(object = "marrayRaw", value = "matrix"): slot assignment method.

print

signature(x = "marrayRaw"): print method for "marrayRaw" class.

Author(s)

Sandrine Dudoit, http://www.stat.berkeley.edu/~sandrine.

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

marrayLayout, marrayNorm, marrayInfo.

Examples


# Examples use swirl dataset, for description type ? swirl
require(limma)
data(swirl)

# Object of class marrayRaw for the 4 swirl arrays
swirl

# Object of class marrayLayout
maLayout(swirl)

# Access only the first 100 spots of the third array
swirl[1:100,3]

# Accessor methods -- How many spots on the array
maNspots(swirl)

# Density plot of log-ratios M for third array
plot(density(maM(swirl[,3])))

# Assignment methods -- Replace maNotes slot
maNotes(swirl)
maNotes(swirl)<-"This is a zebrafish microarray"
maNotes(swirl)

[Package marray version 1.18.0 Index]