plot {marray}	R Documentation

Scatter-plots for cDNA microarray spot statistics

Description

The function maPlot or plot produces scatter-plots of microarray spot statistics for the classes "marrayRaw", "marrayNorm". It also allows the user to highlight and annotate subsets of points on the plot, and display fitted curves from robust local regression or other smoothing procedures.

Usage


## S3 method for class 'marrayRaw':
plot(x, xvar = "maA", yvar = "maM", zvar="maPrintTip", lines.func,text.func,legend.func, ...)
## S3 method for class 'marrayNorm':
plot(x, xvar = "maA", yvar = "maM", zvar="maPrintTip", lines.func,text.func,legend.func, ...)
addText(object, xvar="maA", yvar="maM", subset=NULL, labels=as.character(1:length(subset)), ...)
addPoints(object, xvar="maA", yvar="maM", subset=TRUE, ...)
addLines(object, xvar="maA", yvar="maM", zvar="maPrintTip", subset=TRUE, ...)
## S4 method for signature 'marrayRaw':
text(x, xvar = "maA", yvar = "maM", ...)
## S4 method for signature 'marrayNorm':
text(x, xvar = "maA", yvar = "maM", ...)
## S4 method for signature 'marrayRaw':
lines(x, xvar = "maA", yvar = "maM", ...)
## S4 method for signature 'marrayNorm':
lines(x, xvar = "maA", yvar = "maM", ...)
## S4 method for signature 'marrayRaw':
points(x, xvar = "maA", yvar = "maM", ...)
## S4 method for signature 'marrayNorm':
points(x, xvar = "maA", yvar = "maM", ...)

Arguments

x 	 Microarray object of class "marrayRaw", "marrayNorm".

object 	 Microarray object of class "marrayRaw", "marrayNorm".

xvar 	 Name of accessor function for the abscissa spot statistic, typically a slot name for the microarray object x, such as maA.

yvar 	 Name of accessor function for the ordinate spot statistic, typically a slot name for the microarray object x, such as maM.

zvar 	 Name of accessor method for the spot statistic used to stratify the data, typically a slot name for the microarray layout object (see "marrayLayout") such as maPlate or a method such as maPrintTip. If zvar is NULL, the data are not stratified.

lines.func 	 Function for computing and plotting smoothed fits of y as a function of x, separately within values of zvar, e.g. maLoessLines. If lines.func is NULL, no fitting is performed.

text.func 	 Function for highlighting a subset of points, e.g., maText. If text.func is NULL, no points are highlighted.

legend.func 	 Function for adding a legend to the plot, e.g. maLegendLines. If legend.func is NULL, there is no legend.

subset 	 logical vector or numeric values indicating the subset of points to be plotted.

labels 	 One or more character strings or expressions specifying the text to be written.

... 	 Optional graphical parameters, see par.

Details

This function calls the general function maPlot.func, which is not specific to microarray data. If there are more than one array in the batch, the plot is done for the first array, by default. Default graphical parameters are chosen for convenience using the function maDefaultPar (e.g. color palette, axis labels, plot title) but the user has the option to overwrite these parameters at any point.

Author(s)

Jean Yee Hwa Yang

References

S. Dudoit and Y. H. Yang. (2002). Bioconductor R packages for exploratory analysis and normalization of cDNA microarray data. In G. Parmigiani, E. S. Garrett, R. A. Irizarry and S. L. Zeger, editors, The Analysis of Gene Expression Data: Methods and Software, Springer, New York.

See Also

maPlot.func, maDefaultPar, maLoessLines, maLegendLines, maText, plot, lowess, loess, legend.

Examples


# To see the demo type demo(marrayPlots)

# Examples use swirl dataset, for description type ? swirl
data(swirl)

# Pre-normalization MA-plot for the Swirl 93 array, with the lowess fits for 
# individual print-tip-groups. 
# - Default arguments
plot(swirl[,3])

# Lowess fit using all spots
plot(swirl[,3], zvar=NULL, legend.func=NULL)

# Loess fit using all spots
plot(swirl[,3], zvar=NULL, legend.func=maLegendLines(legend="All spots",col="green"), lines.func=maLoessLines(loess.args=list(span=0.3),col="green"))

[Package marray version 1.18.0 Index]