summary-methods {marray}	R Documentation

Printing summary methods for microarray objects

Description

Print methods were defined for the microarray classes, "marrayInfo", "marrayLayout", "marrayRaw", "marrayNorm". These methods produce summaries of the intensity and textual data stored in different classes of microarray objects.

Methods

x = ANY

generic print method

x = marrayLayout

for an object of class "marrayLayout", the method prints main layout parameters such as the number of spots and the dimensions of the spot and grid matrices.

x = marrayInfo

for an object of class "marrayInfo", the method prints the first 10 rows of the "maInfo" and "maLabels" slots.

x = marrayRaw

for an object of class "marrayRaw", the method prints a short description of the microarray layout "maLayout" and the target samples hybridized to the arrays "maTargets", and a summary of the distribution of the log-ratio statistics "maM".

x = marrayNorm

for an object of class "marrayNorm", the method prints a short description of the microarray layout "maLayout" and the target samples hybridized to the arrays "maTargets", and a summary of the distribution of the log-ratio statistics "maM".

[Package marray version 1.18.0 Index]