read.Galfile {marray}	R Documentation

Reading GenePix Gal file

Description

Reading a standard Gal file containing gene information.

Usage


read.Galfile(galfile, path = ".", info.id = c("ID", "Name"),
layout.id =c(Block="Block", Row="Row", Column="Column"),
labels = "ID", notes = "", sep = "\t", skip = NULL,   ncolumns=4, ...)

Arguments

galfile 	 a character string representing the Gal file.

path 	 a character string representing the data directory. By default this is set to the current working directory (".").

info.id 	 the column numbers or names in `fname' that contain the required information.

layout.id 	 the column names in `fname' that specified the printer layout information.

labels 	 the column number in fname which contains the names that the user would like to use to label spots or arrays (e.g. for default titles in maPlot.

notes 	 object of class character, vector of explanatory text

sep 	 the field separator character. Values on each line of the file are separated by this character. The default is to read a tab delimited file.

skip 	 the number of lines of the data file to skip before beginning to read data.

ncolumns 	 an integer representing the number of columns of sub-array (print-tips) on a slides.

... 	 further arguments to scan.

Value

gnames 	 An object of class marrayInfo.

layout 	 An object of class marrayLayout.

Author(s)

Yee Hwa (Jean) Yang

See Also

read.marrayInfo, read.marrayLayout

Examples


library(marray)
datadir <- system.file("swirldata", package="marray")
try <- read.Galfile(galfile="fish.gal", path=datadir)
names(try)
try$layout
try$gnames

[Package marray version 1.18.0 Index]