read.marrayInfo {marray}	R Documentation

Create objects of class marrayInfo

Description

This function creates objects of class marrayInfo. The marrayInfo class is used to store information regarding the target mRNA samples co-hybridized on the arrays or the spotted probe sequences (e.g. data frame of gene names, annotations, and other identifiers).

Usage


read.marrayInfo(fname, info.id=NULL, labels=NULL, notes=fname, sep="\t",skip=0, quote="\"", ...)

Arguments

fname 	 the name of the file that stores information on target samples or probe sequences. This is usually a file obtained from a database.

info.id 	 the column numbers in fname that contain the required information.

labels 	 the column number in fname which contains the names that the user would like to use to label spots or arrays (e.g. for default titles in maPlot.

notes 	 object of class character, vector of explanatory text

sep 	 the field separator character. Values on each line of the file are separated by this character. The default is to read a tab delimited file.

skip 	 the number of lines of the data file to skip before beginning to read data.

quote 	 the set of quoting characters. By default, this is disable by setting `quote="""'.

... 	 further arguments to scan.

Value

An object of class marrayInfo.

Author(s)

Jean Yang, yeehwa@stat.berkeley.edu

References

http://www.bioconductor.org/

Examples


datadir <- system.file("swirldata", package="marray")

## Reading target sample information
swirl.samples <- read.marrayInfo(file.path(datadir, "SwirlSample.txt"))

## Reading probe sequence information
swirl.gnames <- read.marrayInfo(file.path(datadir, "fish.gal"),
                                info.id=4:5, labels=5, skip=21)

[Package marray version 1.18.0 Index]