read.marrayLayout {marray}	R Documentation

Create objects of class marrayLayout

Description

This function creates objects of class marrayLayout to store layout parameters for two-color cDNA microarrays.

Usage


read.marrayLayout(fname = NULL, ngr, ngc, nsr, nsc, pl.col = NULL, ctl.col = NULL, sub.col = NULL, notes = fname, skip, sep = "\t", quote = "\"", ...)

Arguments

fname 	 the name of the file that stores plate and control information. This is usually a file obtained from a database.

ngr 	 the number of rows of grids per image.

ngc 	 the number of columns of grids per image.

nsr 	 the number of rows of spots per grid.

nsc 	 the number of columns of spots per grid.

pl.col 	 the column number in fname that contains plate information.

ctl.col 	 the column number in fname that contains control information.

sub.col 	 the column number in fname that contains full ID information.

notes 	 object of class character, vector of explanatory text.

skip 	 the number of lines of the data file to skip before beginning to read data.

sep 	 the field separator character. Values on each line of the file are separated by this character. The default is to read a tab delimited file.

quote 	 the set of quoting characters. By default, this is disable by setting `quote="""'.

... 	 further arguments to scan.

Value

An object of class marrayLayout.

Author(s)

Jean Yang yeehwa@stat.berkeley.edu

References

http://www.bioconductor.org/

Examples


datadir <- system.file("swirldata", package="marray")

### Reading in control information from file
skip <-  grep("Row", readLines(file.path(datadir,"fish.gal"), n=100)) - 1
swirl.layout <- read.marrayLayout(fname=file.path(datadir,"fish.gal"), ngr=4, ngc=4,
nsr=22, nsc=24, ctl.col=4, skip=skip)

### Setting control information.
swirl.gnames <- read.marrayInfo(file.path(datadir,"fish.gal"), info.id=4:5, labels=5, skip=21)
x <-  maInfo(swirl.gnames)[,1]
y <- rep(0, maNspots(swirl.layout))
y[x == "control"] <- 1
slot(swirl.layout, "maControls") <- as.factor(y)

[Package marray version 1.18.0 Index]