read.marrayRaw {marray}	R Documentation

Create objects of class "marrayRaw"

Description

This function reads in cDNA microarray data from a directory and creates objects of class "marrayRaw" from spot quantification data files obtained from image analysis software or databases.

Usage


read.marrayRaw(fnames, path=".", name.Gf=NULL, name.Gb=NULL, name.Rf=NULL,
name.Rb=NULL,name.W=NULL, layout=NULL, gnames=NULL, targets=NULL,
notes=NULL, skip=NULL, sep="    ", quote="\"", DEBUG=FALSE, ...)

read.GenePix(fnames = NULL, path = NULL, name.Gf = "F532 Median",
name.Gb ="B532 Median", name.Rf = "F635 Median", name.Rb = "B635 Median",
name.W ="Flags", layout = NULL, gnames = NULL, targets = NULL,
notes = NULL, skip=NULL, sep = "        ", quote = "\"", DEBUG=FALSE, ...)

read.SMD(fnames = NULL, path = NULL, name.Gf = "Ch1 Intensity (Median)",
name.Gb = "Ch1 Background (Median)", name.Rf = "Ch2 Intensity (Median)",
name.Rb = "Ch2 Background (Median)", name.W = NULL, info.id = c("Name",
"Clone ID"), layout = NULL, gnames = NULL, targets = NULL, notes = NULL, skip = NULL, sep = "\t", quote = "\"", DEBUG=FALSE, ...)

read.Spot(fnames = NULL, path = ".", name.Gf = "Gmean", name.Gb =
"morphG", name.Rf = "Rmean", name.Rb = "morphR",name.W = NULL, layout =
NULL, gnames = NULL, targets = NULL, notes = NULL, skip = NULL, sep = "\t", quote = "\"", DEBUG=FALSE, ...)

read.Agilent(fnames = NULL, path=NULL, name.Gf = "gMedianSignal", name.Gb = "gBGMedianSignal", name.Rf = "rMedianSignal", name.Rb = "rBGMedianSignal", name.W= NULL, layout = NULL, gnames = NULL, targets = NULL, notes=NULL, skip=NULL, sep="\t", quote="\"", DEBUG=FALSE, info.id=NULL, ...)

widget.marrayRaw(ext = c("spot", "xls", "gpr"), skip = 0, sep = "\t",  quote = "\"", ...)

Arguments

fnames 	 a vector of character strings containing the file names of each spot quantification data file. These typically end in .spot for the software Spot or .gpr for the software GenePix.

path 	 a character string representing the data directory. By default this is set to the current working directory ("."). In the case where fnames contains the full path name, path should be set to NULL.

name.Gf 	 character string for the column header for green foreground intensities.

name.Gb 	 character string for the column header for green background intensities.

name.Rf 	 character string for the column header for red foreground intensities.

name.Rb 	 character string for the column header for red background intensities.

name.W 	 character string for the column header for spot quality weights.

layout 	 object of class "marrayLayout", containing microarray layout parameters.

gnames 	 object of class "marrayInfo" containing probe sequence information.

targets 	 object of class "marrayInfo" containing target sample information.

notes 	 object of class "character", vector of explanatory text.

info.id 	 object of class "character", vector containing the name of the colums of the SMD file containing oligo information you want to retrieve. By default, this is set to read Homo sapiens data. You may need to modify this argument if your are working on another genome.

skip 	 the number of lines of the data file to skip before beginning to read in data.

sep 	 the field separator character. Values on each line of the file are separated by this character. The default is to read a tab delimited file.

quote 	 the set of quoting characters. By default, this is disabled by setting quote="\"".

ext 	 a characters string representing suffix of different image analysis output files.

DEBUG 	 a logical value, if TRUE, a series of echo statements will be printed.

... 	 further arguments to scan.

Value

An object of class "marrayRaw".

Author(s)

Jean Yang, yeehwa@stat.berkeley.edu

References

http://www.bioconductor.org/.

See Also

scan, read.marrayLayout, read.marrayInfo

Examples


datadir <- system.file("swirldata", package="marray")

## Quick guide
swirl.targets <- read.marrayInfo(file.path(datadir, "SwirlSample.txt"))
data <- read.Spot(path=datadir, targets=swirl.targets)

## Alternate commands
skip <-  grep("Row", readLines(file.path(datadir,"fish.gal"), n=100)) - 1

swirl.layout <- read.marrayLayout(ngr=4, ngc=4, nsr=22, nsc=24)

swirl.targets <- read.marrayInfo(file.path(datadir, "SwirlSample.txt"))

swirl.gnames <- read.marrayInfo(file.path(datadir, "fish.gal"),
                                info.id=4:5, labels=5, skip=skip)

x <-  maInfo(swirl.gnames)[,1]
y <- rep(0, maNspots(swirl.layout))
y[x == "control"] <- 1
slot(swirl.layout, "maControls") <- as.factor(y)

fnames <- dir(path=datadir,pattern="spot")
swirl<- read.Spot(fnames, path=datadir,
                       layout = swirl.layout,
                       gnames = swirl.gnames,
                       targets = swirl.targets)


[Package marray version 1.18.0 Index]