ShowLargeObject-class {marray}	R Documentation

Show Large Data Object - class

Description

A virtual class including the data classes marrayRaw, marrayNorm, marrayInfo, marrayLayout, PrinterInfo, RGData and MAData, all of which typically contain large quantities of numerical data in vector, matrices and data.frames.

Methods

A show method is defined for objects of class ShowLargeObject which uses printHead to print only the leading elements or rows of components or slots which contain large quantities of data.

Author(s)

modifid from Gordon Smyth's function

[Package marray version 1.18.0 Index]