stat.confband.text {marray}	R Documentation

Rank genes according to the value of a statistic.

Description

Select values based on intensities binning.

Usage


stat.confband.text(M, A, crit1=0.025, crit2=crit1, nclass=5)

Arguments

A 	 a vector giving the x-coordinates of the points in the scatter plot. In the microarray context, this could be a vector of average log intensities. ie A

M 	 a vector giving the y-coordinates of the points in the scatter plot. In the microarray context, this could be a vector of log intensity ratios.

crit1 	 The number of points to be selected. If crit1 < 1, the crit1*100% spots with the smallest M values will be selected. If crit1 >= 1, the crit spots with the smallest M values are selected.

crit2 	 Similar to "crit1". If crit2 < 1, the crit2*100% spots with the largest M values will be selected. If crit2 >= 1, the crit2 spots with the largest M values are selected.

nclass 	 A single number giving the approximate number of intensity depedent groups to consider.

Value

A vector of selected spot index.

See Also

stat.gnames

Examples


library(marray)
data(swirl)
aveA <- apply(maA(swirl), 1, mean.na)
aveM <- apply(maM(swirl), 1, mean.na)
stat.confband.text(aveM, aveA, crit1=20, crit2=50, nclass=5)

[Package marray version 1.18.0 Index]