stat.gnames {marray}	R Documentation

Sort Genes According to the Value of a Statistic

Description

Lists genes and corresponding statistics in decreasing order of the statistics. This function applies to any type of statistic, including log ratios, one and two-sample t-statistics, and F-statistics. Missing values are ignored, as in sort(..., na.last=NA).

Usage


stat.gnames(x, gnames, crit= 50)

Arguments

x 	 a numeric vector containing the statistics for each gene. Missing values (NAs) are allowed.

gnames 	 a character vector containing the gene names.

crit 	 specifies the number of genes to be returned. If crit < 1, the crit*100% genes with the largest x values are listed. If crit >= 1, the crit genes with the largest x values are listed.

Value

List containing the following components

gnames 	 gene names sorted in decreasing order of the statistics in x.

t 	 statistics sorted in decreasing order.

Author(s)

Yee Hwa Yang, yeehwa@stat.berkeley.edu
Sandrine Dudoit, sandrine@stat.berkeley.edu

See Also

order, sort.

Examples


data(swirl)
aveM <- apply(maM(swirl), 1, mean.na)
Gnames <- maGeneTable(swirl)

stat.gnames(abs(aveM), Gnames, crit=10)
stat.gnames(aveM, Gnames, crit=0.01)


[Package marray version 1.18.0 Index]