write.list {marray}	R Documentation

Data Output

Description

Writes information from a list into a text file.

Usage


write.list(x, filename = "data", append = FALSE, closefile = TRUE, outfile)

Arguments

x 	 the list object to be written.

filename 	 a character string representing the file name.

append 	 logical; if true, the data x is appended to file filename.

closefile 	 logical indicating if the file connection should be closed.

outfile 	 file name or connections.

Details

This function may be called recursively if there exists list structure within a list.

Author(s)

Jean Yee Hwa Yang

See Also

write.table, write

Examples


data(swirl)
test <- list(A = 1:10, B= maM(swirl)[1:10,], C=list(x=1:10, y=1:4),
             D = summary(maA(swirl[,1])))
write.list(test, filename="test.txt")

[Package marray version 1.18.0 Index]