write.xls {marray}	R Documentation

Data Output

Description

Calls the function write.table with predefine argument. The entries in each line (row) are separated by tab.

Usage


write.xls(res, file = "test.xls", ...)

Arguments

res 	 the object to be written, typically a data frame. If not, it is attempted to coerce x to a data frame.

file 	 a character string representing the file name.

... 	 further arguments to write.table.

Details

see write.table

Author(s)

Jean Yee Hwa Yang

See Also

write.table, write.list

Examples


data(swirl)
write.xls(maM(swirl)[1:10,], "normM.xls")

[Package marray version 1.18.0 Index]